+++ src/dgbowl_schemas/yadg_dataschema/dataschema_4_0/__init__.py
@@ -1,0 +1,1 @@
+from .schema import DataSchema
