+++ src/dgbowl_schemas/yadg_dataschema/dataschema_4_0/externaldate.py
@@ -30,32 +30,16 @@ class ExternalDateUTSOffset(BaseModel, extra=Extra.forbid):
     utsoffset: float
 
 
-class ExternalDate(BaseModel, extra=Extra.forbid):
+class ExternalDate(BaseModel, extra=Extra.forbid, allow_population_by_field_name=True):
     using: Union[
         ExternalDateFile,
         ExternalDateFilename,
         ExternalDateISOString,
         ExternalDateUTSOffset,
-    ]
+    ] = Field(default=None, alias="from")
     mode: Literal["add", "replace"] = "add"
     _default: bool = PrivateAttr()
 
     def __init__(self, **data):
         self._default = False
         super().__init__(**data)
-
-    @root_validator(pre=True, allow_reuse=True)
-    def check_using(cls, values):
-        using = values.pop("using", None)
-        if using is None and "from" in values:
-            logger.warning(
-                "Specifying 'from' for ExternalDate was deprecated in "
-                "dgbowl_schemas-v103, and may stop working in future "
-                "versions of DataSchema. Please use 'using' instead."
-            )
-            using = values.pop("from")
-        if using is None:
-            using = {"filename": {"format": "%Y-%m-%d-%H-%M-%S", "len": 19}}
-            cls._default = True
-        values["using"] = using
-        return values
