+++ src/dgbowl_schemas/yadg_dataschema/dataschema_4_0/schema.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel, Extra
+from typing import Sequence
+from .metadata import Metadata
+from .step import Step
+
+
+class DataSchema(BaseModel, extra=Extra.forbid):
+    metadata: Metadata
+    steps: Sequence[Step]
