+++ src/dgbowl_schemas/yadg_dataschema/schema.py
@@ -1,94 +1,6 @@
-from pydantic import BaseModel, Field, validator, root_validator, Extra
-from typing import Optional, Union, Sequence, Mapping, Literal
-import logging
+#from typing import Union
+#from .dataschema_4_1 import DataSchema as DataSchema_4_1
+from .dataschema_4_0 import DataSchema as DataSchema_4_0
 
-logger = logging.getLogger(__name__)
-
-from .parameters import (
-    Dummy,
-    BasicCSV,
-    MeasCSV,
-    XPSTrace,
-    XRDTrace,
-    ChromTrace,
-    ElectroChem,
-    FlowData,
-    MassTrace,
-    QFTrace,
-)
-from .externaldate import ExternalDate
-from .input import Input
-
-
-class Metadata(BaseModel, extra=Extra.forbid):
-    provenance: str
-    version: str
-    timezone: str = "localtime"
-    provenance_metadata: Optional[Mapping]
-
-    @root_validator(pre=True, allow_reuse=True)
-    def check_version(cls, values):
-        version = values.get("version")
-        if version is None:
-            for k in {"dataschema_version", "schema_version"}:
-                if k in values:
-                    version = values.pop(k)
-        values["version"] = version
-        return values
-
-
-class Step(BaseModel, extra=Extra.forbid):
-    parser: Literal[
-        "dummy",
-        "basiccsv",
-        "meascsv",
-        "xpstrace",
-        "xrdtrace",
-        "chromtrace",
-        "electrochem",
-        "flowdata",
-        "masstrace",
-        "qftrace",
-    ]
-    input: Input
-    tag: str = None
-    parameters: Union[
-        Dummy,
-        BasicCSV,
-        MeasCSV,
-        XPSTrace,
-        XRDTrace,
-        ChromTrace,
-        ElectroChem,
-        FlowData,
-        MassTrace,
-        QFTrace,
-    ] = Field(default=None, discriminator="parser")
-    externaldate: Optional[ExternalDate]
-    export: Optional[str]
-
-    @root_validator(pre=True, allow_reuse=True)
-    def check_import(cls, values):
-        input = values.get("input")
-        if input is None and "import" in values:
-            logger.warning(
-                "Specifying 'input' files of a Step using the 'import' key is "
-                "deprecated and may stop working in future versions of DataSchema."
-            )
-            input = values.pop("import")
-        if input is None:
-            raise ValueError("The 'input' section has to be specified for each Step.")
-        values["input"] = input
-        return values
-
-    @root_validator(pre=True, allow_reuse=True)
-    def copy_parser_to_params(cls, values):
-        params = values.get("parameters")
-        if params is not None:
-            values["parameters"]["parser"] = values["parser"]
-        return values
-
-
-class DataSchema(BaseModel, extra=Extra.forbid):
-    metadata: Metadata
-    steps: Sequence[Step]
+#DataSchema = Union[DataSchema_4_1, DataSchema_4_0]
+DataSchema = DataSchema_4_0
