+++ tests/test_dataschema/ts0_dummy.json
@@ -1,7 +1,7 @@
 {
     "metadata": {
         "provenance": "manual",
-        "schema_version": "0.1"
+        "schema_version": "4.0"
     },
     "steps": [
         {
